+++ app/src/main/java/org/secfirst/umbrella/whitelabel/feature/segment/view/SegmentController.kt
@@ -1,22 +1,12 @@
 package org.secfirst.umbrella.whitelabel.feature.segment.view
 
-import android.app.Dialog
-import android.content.Intent
-import android.net.Uri
 import android.os.Bundle
-import android.support.v4.app.ShareCompat
-import android.support.v4.content.FileProvider
-import android.support.v7.widget.AppCompatButton
 import android.support.v7.widget.GridLayoutManager
 import android.view.*
-import android.widget.RadioButton
-import android.widget.RadioGroup
 import com.bluelinelabs.conductor.RouterTransaction
 import com.commonsware.cwac.anddown.AndDown
 import kotlinx.android.synthetic.main.segment_view.*
-import org.apache.commons.io.FilenameUtils.removeExtension
 import org.jsoup.Jsoup
-import org.secfirst.umbrella.whitelabel.BuildConfig
 import org.secfirst.umbrella.whitelabel.R
 import org.secfirst.umbrella.whitelabel.UmbrellaApplication
 import org.secfirst.umbrella.whitelabel.data.database.checklist.Checklist
@@ -28,10 +18,8 @@ import org.secfirst.umbrella.whitelabel.feature.segment.DaggerSegmentComponent
 import org.secfirst.umbrella.whitelabel.feature.segment.interactor.SegmentBaseInteractor
 import org.secfirst.umbrella.whitelabel.feature.segment.presenter.SegmentBasePresenter
 import org.secfirst.umbrella.whitelabel.feature.segment.view.adapter.SegmentAdapter
-import org.secfirst.umbrella.whitelabel.misc.FileExtensions
-import org.secfirst.umbrella.whitelabel.misc.createDocument
 import org.secfirst.umbrella.whitelabel.misc.initGridView
-import java.io.File
+import org.secfirst.umbrella.whitelabel.misc.showShareDialog
 import javax.inject.Inject
 
 
@@ -126,10 +114,9 @@ class SegmentController(bundle: Bundle) : BaseController(bundle), SegmentView {
 
     private fun onChecklistShareClick() {
         val checklistHtml = checklist?.covertToHTML()
-        val intent = Intent(Intent.ACTION_SENDTO,
-                Uri.parse("mailto:?subject=Checklist&body=${Uri.encode(checklistHtml)}"))
-        startActivity(intent)
-
+        val doc = Jsoup.parse(checklistHtml)
+        doc.outputSettings().syntax(org.jsoup.nodes.Document.OutputSettings.Syntax.xml)
+        showShareDialog(doc, "Shared Checklist", context, activity!!)
     }
 
     private fun onSegmentShareClick(markdown: Markdown) {
@@ -137,7 +124,7 @@ class SegmentController(bundle: Bundle) : BaseController(bundle), SegmentView {
         val result = andDown.markdownToHtml(markdown.text, AndDown.HOEDOWN_EXT_QUOTE, 0)
         val doc = Jsoup.parse(result)
         doc.outputSettings().syntax(org.jsoup.nodes.Document.OutputSettings.Syntax.xml)
-        showShareDialog(doc, markdown.title)
+        showShareDialog(doc, markdown.title, context, activity!!)
     }
 
     private fun onFootClicked(position: Int) {
@@ -158,69 +145,5 @@ class SegmentController(bundle: Bundle) : BaseController(bundle), SegmentView {
 
     fun setIndexTab(position: Int) {
         this.indexTab = position
-    }
-
-
-    private fun shareDocument(fileToShare: File) {
-
-        val uri = FileProvider.getUriForFile(context, BuildConfig.APPLICATION_ID, fileToShare)
-        val shareIntent = ShareCompat.IntentBuilder.from(activity)
-                .setType(activity!!.contentResolver.getType(uri))
-                .setStream(uri)
-                .intent
-
-        //Provide read access
-        shareIntent.action = Intent.ACTION_SEND
-        shareIntent.putExtra(Intent.EXTRA_SUBJECT, removeExtension(fileToShare.name))
-        shareIntent.addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-        val pm = activity!!.packageManager
-
-        if (shareIntent.resolveActivity(pm) != null) {
-            startActivity(Intent.createChooser(shareIntent, R.string.export_lesson.toString()))
-        }
-
-    }
-
-
-    //Share Menu
-    private fun showShareDialog(doc: org.jsoup.nodes.Document, title: String) {
-
-        var type = FileExtensions.PDF
-        // custom dialog
-        val dialog = Dialog(context)
-        dialog.setContentView(R.layout.share_dialog)
-
-        val shareWindow: RadioGroup = dialog.findViewById(R.id.radio_group)
-
-        for (i in 0 until FileExtensions.values().size) {
-            val radioButton = RadioButton(context)
-            radioButton.text = FileExtensions.values()[i].toString()
-            shareWindow.addView(radioButton)
-        }
-
-        shareWindow.check(shareWindow.getChildAt(0).id)
-
-        val shareButton: AppCompatButton = dialog.findViewById(R.id.share_document_button)
-        shareButton.setOnClickListener { _ ->
-            shareDocument(createDocument(doc, title, type, context))
-            dialog.dismiss()
-        }
-
-        val dismissButton: AppCompatButton = dialog.findViewById(R.id.cancel_share_button)
-        dismissButton.setOnClickListener { _ ->
-            dialog.dismiss()
-        }
-
-        dialog.show();
-
-        shareWindow.setOnCheckedChangeListener(RadioGroup.OnCheckedChangeListener { group, checkedId ->
-            val childCount = group.childCount
-            for (x in 0 until childCount) {
-                val btn = group.getChildAt(x) as RadioButton
-                if (btn.id == checkedId) {
-                    type = FileExtensions.valueOf(btn.text.toString())
-                }
-            }
-        })
     }
 }
